renovate webpack config
