+++ src/client/index.jsx
@@ -1,11 +1,26 @@
 import React from 'react';
-import { render } from 'react-dom';
+import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
-import App from './app';
+import { AppContainer } from 'react-hot-loader';
+import App from './client/app';
 
-render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
-  document.getElementById('app')
-);
+const render = (Component) => {
+  ReactDOM.render(
+    <AppContainer>
+      <BrowserRouter>
+        <Component />
+      </BrowserRouter>
+    </AppContainer>,
+    document.getElementById('root')
+  );
+};
+
+render(App);
+
+// Hot Module Replacement API
+if (module.hot) {
+  module.hot.accept('./client/app', () => {
+    render(App);
+  });
+}
+
